+++ src/_h5ai/public/js/lib/ext/preview/preview.js
@@ -213,7 +213,7 @@ Session.prototype = {
                     }
                 });
                 dom('#pv-container').hide().clr();
-                showSpinner(true, item.thumbSquare || item.icon, 200);
+                showSpinner(true, item.thumbRational || item.icon, 200);
             })
             .then(() => this.load(item))
             // delay for testing
@@ -235,7 +235,7 @@ Session.prototype = {
 const register = (types, load, adjust) => {
     const initItem = item => {
         if (item.$view && includes(types, item.type)) {
-            item.$view.find('a').on('click', ev => {
+            const onclick = ev => {
                 ev.preventDefault();
 
                 const matchedItems = compact(dom('#items .item').map(el => {
@@ -245,11 +245,22 @@ const register = (types, load, adjust) => {
 
                 session = Session(matchedItems, matchedItems.indexOf(item), load, adjust);
                 enter();
-            });
+            }
+
+            if (item.click_callback) {
+                item.$view.find('a').off('click', item.click_callback);
+            }
+            item.click_callback = onclick;
+            item.click_callback.type = item.type;
+            item.$view.find('a').on('click', onclick);
+        }
+        else if (item.$view && item.click_callback && includes(types, item.click_callback.type)) {
+            item.$view.find('a').off('click', item.click_callback);
         }
     };
 
     event.sub('view.changed', added => each(added, initItem));
+    event.sub('item.changed', changed => initItem(changed));
 };
 
 const init = () => {